// src/app/business-logic/model/organization/organizationCreateInviteResponse.ts
/**
 * organization
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface OrganizationCreateInviteResponse {
    /**
     * Invite ID
     */
    id?: string;
    /**
     * Maximum allowed users in the company
     */
    allowed_users?: number;
    /**
     * The users already allocated to the company
     */
    allocated_users?: number;
}