// src/app/business-logic/model/users/usersGetCurrentUserResponseTermsOfUse.ts
/**
 * users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.3
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TermsOfUse } from './termsOfUse';


/**
 * Terms of use information
 */
export interface UsersGetCurrentUserResponseTermsOfUse {
    /**
     * User is required to accept new terms of use (terms of use were never accepted or accepted version has been deprecated)
     */
    accept_required?: boolean;
    latest?: TermsOfUse;
}